Add files via upload
The decrease in bee populations can be attributed to a variety of factors, including intensified agricultural practices, a rise in pest and disease pressures, the use of chemicals like neonicotinoid pesticides, the effects of climate change, and a decrease in beekeeping efforts. 

This project focuses on the decline of honey production and uses past trends to predict the future of honeybees and their production.